refactor : float, animation components
