+++ src/components/animation/snowflake/index.scss
